refactor(github-delivery): share event fixtures
