+++ crates/automata-ci-github-delivery/tests/delivery_service.rs
@@ -14,12 +14,6 @@ use automata_ci_blob::{
     MediaType, PutBlobOutcome, VerifiedBlob,
 };
 use automata_ci_core::{Sha256Digest, UnixMillis};
-use automata_ci_github::{
-    GITHUB_EVENT_ENVELOPE_V1_MEDIA_TYPE,
-    GithubRepositoryVisibility as GithubRepositoryVisibilityFact, GithubSealedEventEnvelopeV1,
-    GithubWebhookBodyDigest, StoredAuthenticatedGithubWebhook,
-    rehydrate_stored_authenticated_github_webhook,
-};
 use automata_ci_github_delivery::{
     GITHUB_AUTHENTICATED_EVENT_MEDIA_TYPE, GithubChangedFileSelection,
     GithubChangedFilesDisposition, GithubDeliveryClock, GithubDeliveryPrivateRepositoryAction,
@@ -64,9 +58,7 @@ use automata_ci_store::{
 };
 use automata_ci_workflow_service::{GithubWorkflowPlanVerifier, WorkflowAdmissionService};
 use bytes::Bytes;
-use flate2::{Compression, write::GzEncoder};
 use sha2::{Digest as _, Sha256};
-use tar::{Builder, EntryType, Header};
 use tokio::sync::Notify;
 use tokio_util::sync::CancellationToken;
 use uuid::Uuid;
@@ -74,15 +66,12 @@ use uuid::Uuid;
 use super::subject_evidence::{
     fixture_check_head_sha, fixture_subject_evidence, fixture_subject_evidence_with_head,
 };
+use super::support::{
+    AFTER, BEFORE, INSTALLATION_ID, OWNER, REPOSITORY, REPOSITORY_ID, REPOSITORY_OWNER_ID, ZERO,
+    archive, provider_event_envelope, push_body,
+};
 
-const BEFORE: &str = "fedcba9876543210fedcba9876543210fedcba98";
-const AFTER: &str = "0123456789abcdef0123456789abcdef01234567";
-const ZERO: &str = "0000000000000000000000000000000000000000";
-const OWNER: &str = "octo-private";
-const REPOSITORY: &str = "private-repository";
-const REPOSITORY_ID: u64 = 9_001;
-const REPOSITORY_OWNER_ID: u64 = 8_001;
-const INSTALLATION_ID: u64 = 4_242;
+const DELIVERY: &str = "delivery-service-1";
 const TOKEN_MARKER: &str = "delivery-service-token-marker";
 const INITIAL_NOW: i64 = 100;
 const CLAIM_MILLIS: i64 = 50;
@@ -1813,7 +1802,8 @@ fn delivery_template(
     stored_visibility: ProviderRepositoryVisibility,
     deleted: bool,
 ) -> (DeliveryTemplate, BlobDescriptor, Bytes) {
-    let body = push_body(stored_visibility, deleted);
+    let after = if deleted { ZERO } else { AFTER };
+    let body = push_body("refs/heads/main", after, deleted, 0, stored_visibility);
     let digest = Sha256Digest::from_bytes(Sha256::digest(&body).into());
     let key_text = format!("provider-deliveries/github/event/sha256/{digest}.json");
     let descriptor = BlobDescriptor::new(
@@ -1829,7 +1819,8 @@ fn delivery_template(
         GITHUB_AUTHENTICATED_EVENT_MEDIA_TYPE,
     )
     .expect("raw event");
-    let event_envelope = provider_event_envelope(&body, &descriptor, stored_visibility);
+    let event_envelope =
+        provider_event_envelope(&body, &descriptor, "push", DELIVERY, stored_visibility);
     let delivery_id = ProviderDeliveryId::from_uuid(Uuid::from_u128(1)).expect("delivery ID");
     let repository = ProviderRepositoryCoordinates::new(
         ProviderRepositoryId::new(REPOSITORY_ID).expect("repository"),
@@ -1843,7 +1834,7 @@ fn delivery_template(
         ProviderConnectionId::from_uuid(Uuid::from_u128(3)).expect("connection"),
         ProviderInstallationId::new(INSTALLATION_ID).expect("installation"),
         repository,
-        "delivery-service-1",
+        DELIVERY,
     )
     .expect("delivery identity");
     (
@@ -1860,54 +1851,6 @@ fn delivery_template(
     )
 }
 
-fn provider_event_envelope(
-    body: &Bytes,
-    descriptor: &BlobDescriptor,
-    visibility: ProviderRepositoryVisibility,
-) -> ProviderDeliveryEventEnvelope {
-    let visibility = match visibility {
-        ProviderRepositoryVisibility::Public => GithubRepositoryVisibilityFact::Public,
-        ProviderRepositoryVisibility::Private => GithubRepositoryVisibilityFact::Private,
-    };
-    let stored = StoredAuthenticatedGithubWebhook::from_durable_coordinates(
-        body.clone(),
-        GithubWebhookBodyDigest::from_bytes(*descriptor.digest().as_bytes()),
-        descriptor.size(),
-        GITHUB_AUTHENTICATED_EVENT_MEDIA_TYPE,
-        "push",
-        "delivery-service-1",
-        INSTALLATION_ID,
-        REPOSITORY_ID,
-        REPOSITORY_OWNER_ID,
-        visibility,
-        OWNER,
-        REPOSITORY,
-    );
-    let event =
-        rehydrate_stored_authenticated_github_webhook(stored).expect("verified webhook fixture");
-    let sealed = GithubSealedEventEnvelopeV1::seal(&event, descriptor.clone())
-        .expect("sealed event envelope fixture");
-    ProviderDeliveryEventEnvelope::new(
-        sealed.schema(),
-        sealed.registry_schema(),
-        sealed.digest(),
-        sealed.canonical_bytes().to_vec(),
-        GITHUB_EVENT_ENVELOPE_V1_MEDIA_TYPE,
-    )
-    .expect("durable event envelope fixture")
-}
-
-fn push_body(visibility: ProviderRepositoryVisibility, deleted: bool) -> Bytes {
-    let (private, visibility) = match visibility {
-        ProviderRepositoryVisibility::Public => (false, "public"),
-        ProviderRepositoryVisibility::Private => (true, "private"),
-    };
-    let after = if deleted { ZERO } else { AFTER };
-    Bytes::from(format!(
-        r#"{{"ref":"refs/heads/main","before":"{BEFORE}","after":"{after}","created":false,"deleted":{deleted},"forced":false,"repository":{{"id":{REPOSITORY_ID},"private":{private},"visibility":"{visibility}","name":"{REPOSITORY}","full_name":"{OWNER}/{REPOSITORY}","owner":{{"id":{REPOSITORY_OWNER_ID},"login":"{OWNER}"}}}},"installation":{{"id":{INSTALLATION_ID}}},"commits":[]}}"#,
-    ))
-}
-
 fn repository_source() -> RepositorySource {
     RepositorySource::from_bytes(
         ScmProviderId::new("github").expect("provider"),
@@ -1919,41 +1862,6 @@ fn repository_source() -> RepositorySource {
             b"on: push\n".to_vec(),
         )])),
     )
-}
-
-fn archive(files: BTreeMap<&str, Vec<u8>>) -> Bytes {
-    let encoder = GzEncoder::new(Vec::new(), Compression::default());
-    let mut builder = Builder::new(encoder);
-    append_archive_entry(&mut builder, "repository-root", EntryType::Directory, &[]);
-    for (path, bytes) in files {
-        append_archive_entry(
-            &mut builder,
-            &format!("repository-root/{path}"),
-            EntryType::Regular,
-            &bytes,
-        );
-    }
-    let encoder = builder.into_inner().expect("finish tar");
-    Bytes::from(encoder.finish().expect("finish gzip"))
-}
-
-fn append_archive_entry(
-    builder: &mut Builder<GzEncoder<Vec<u8>>>,
-    path: &str,
-    entry_type: EntryType,
-    bytes: &[u8],
-) {
-    let mut header = Header::new_gnu();
-    header.set_entry_type(entry_type);
-    header.set_mode(if entry_type.is_dir() { 0o755 } else { 0o644 });
-    header.set_uid(0);
-    header.set_gid(0);
-    header.set_mtime(0);
-    header.set_size(u64::try_from(bytes.len()).expect("entry size"));
-    header.set_cksum();
-    builder
-        .append_data(&mut header, path, bytes)
-        .expect("append archive entry");
 }
 
 async fn run_once(
