+++ crates/automata-ci-github-delivery/tests/github_delivery.rs
@@ -1,4 +1,5 @@
 mod subject_evidence;
+mod support;
 
 mod checks_publisher;
 mod delivery_service;
