+++ crates/automata-ci-github-delivery/tests/support/mod.rs
@@ -1,0 +1,117 @@
+use std::collections::BTreeMap;
+
+use automata_ci_blob::BlobDescriptor;
+use automata_ci_github::{
+    GITHUB_EVENT_ENVELOPE_V1_MEDIA_TYPE,
+    GithubRepositoryVisibility as GithubRepositoryVisibilityFact, GithubSealedEventEnvelopeV1,
+    GithubWebhookBodyDigest, StoredAuthenticatedGithubWebhook,
+    rehydrate_stored_authenticated_github_webhook,
+};
+use automata_ci_github_delivery::GITHUB_AUTHENTICATED_EVENT_MEDIA_TYPE;
+use automata_ci_store::{ProviderDeliveryEventEnvelope, ProviderRepositoryVisibility};
+use bytes::Bytes;
+use flate2::{Compression, write::GzEncoder};
+use tar::{Builder, EntryType, Header};
+
+pub(super) const BEFORE: &str = "fedcba9876543210fedcba9876543210fedcba98";
+pub(super) const AFTER: &str = "0123456789abcdef0123456789abcdef01234567";
+pub(super) const ZERO: &str = "0000000000000000000000000000000000000000";
+pub(super) const OWNER: &str = "octo-private";
+pub(super) const REPOSITORY: &str = "private-repository";
+pub(super) const REPOSITORY_ID: u64 = 9_001;
+pub(super) const REPOSITORY_OWNER_ID: u64 = 8_001;
+pub(super) const INSTALLATION_ID: u64 = 4_242;
+
+pub(super) fn archive<T: AsRef<[u8]>>(files: BTreeMap<&str, T>) -> Bytes {
+    let encoder = GzEncoder::new(Vec::new(), Compression::default());
+    let mut builder = Builder::new(encoder);
+    append_archive_entry(&mut builder, "repository-root", EntryType::Directory, &[]);
+    for (path, bytes) in files {
+        append_archive_entry(
+            &mut builder,
+            &format!("repository-root/{path}"),
+            EntryType::Regular,
+            bytes.as_ref(),
+        );
+    }
+    let encoder = builder.into_inner().expect("finish tar");
+    Bytes::from(encoder.finish().expect("finish gzip"))
+}
+
+fn append_archive_entry(
+    builder: &mut Builder<GzEncoder<Vec<u8>>>,
+    path: &str,
+    entry_type: EntryType,
+    bytes: &[u8],
+) {
+    let mut header = Header::new_gnu();
+    header.set_entry_type(entry_type);
+    header.set_mode(if entry_type.is_dir() { 0o755 } else { 0o644 });
+    header.set_uid(0);
+    header.set_gid(0);
+    header.set_mtime(0);
+    header.set_size(u64::try_from(bytes.len()).expect("entry size"));
+    header.set_cksum();
+    builder
+        .append_data(&mut header, path, bytes)
+        .expect("append archive entry");
+}
+
+pub(super) fn provider_event_envelope(
+    body: &Bytes,
+    descriptor: &BlobDescriptor,
+    event_name: &str,
+    delivery: &str,
+    visibility: ProviderRepositoryVisibility,
+) -> ProviderDeliveryEventEnvelope {
+    let visibility = match visibility {
+        ProviderRepositoryVisibility::Public => GithubRepositoryVisibilityFact::Public,
+        ProviderRepositoryVisibility::Private => GithubRepositoryVisibilityFact::Private,
+    };
+    let stored = StoredAuthenticatedGithubWebhook::from_durable_coordinates(
+        body.clone(),
+        GithubWebhookBodyDigest::from_bytes(*descriptor.digest().as_bytes()),
+        descriptor.size(),
+        GITHUB_AUTHENTICATED_EVENT_MEDIA_TYPE,
+        event_name,
+        delivery,
+        INSTALLATION_ID,
+        REPOSITORY_ID,
+        REPOSITORY_OWNER_ID,
+        visibility,
+        OWNER,
+        REPOSITORY,
+    );
+    let event =
+        rehydrate_stored_authenticated_github_webhook(stored).expect("verified webhook fixture");
+    let sealed = GithubSealedEventEnvelopeV1::seal(&event, descriptor.clone())
+        .expect("sealed event envelope fixture");
+    ProviderDeliveryEventEnvelope::new(
+        sealed.schema(),
+        sealed.registry_schema(),
+        sealed.digest(),
+        sealed.canonical_bytes().to_vec(),
+        GITHUB_EVENT_ENVELOPE_V1_MEDIA_TYPE,
+    )
+    .expect("durable event envelope fixture")
+}
+
+pub(super) fn push_body(
+    git_ref: &str,
+    after: &str,
+    deleted: bool,
+    commit_count: usize,
+    visibility: ProviderRepositoryVisibility,
+) -> Bytes {
+    let commits = (1..=commit_count)
+        .map(|value| format!(r#"{{"id":"{value:040x}"}}"#))
+        .collect::<Vec<_>>()
+        .join(",");
+    let (private, visibility) = match visibility {
+        ProviderRepositoryVisibility::Public => (false, "public"),
+        ProviderRepositoryVisibility::Private => (true, "private"),
+    };
+    Bytes::from(format!(
+        r#"{{"ref":"{git_ref}","before":"{BEFORE}","after":"{after}","created":false,"deleted":{deleted},"forced":false,"repository":{{"id":{REPOSITORY_ID},"private":{private},"visibility":"{visibility}","name":"{REPOSITORY}","full_name":"{OWNER}/{REPOSITORY}","owner":{{"id":{REPOSITORY_OWNER_ID},"login":"{OWNER}"}}}},"installation":{{"id":{INSTALLATION_ID}}},"commits":[{commits}]}}"#,
+    ))
+}
